+++ README.md
@@ -1,1 +1,9 @@
-# backEnd
+# Flask API
+#### Integração entre banco de dados e Arduino com Flask e SQLAlchemy (Python)
+
+ ---
+
+Esta API fornecerá rotas a serem acessadas remotamente por um módulo Arduino, com o intuito de obter informações sobre a autorização de alunos e professores aos laboratórios do Instituto de Computação da UFBA.
+
+<p align="center">
+<img src="https://github.com/igordantasgf/igordantasgf/blob/main/Untitled%20-%20Frame%201%20(1).jpg" alt="arquiteturaapi" height="400" style="vertical-align:top; margin:4px">
