+++ main/aluno/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+aluno = Blueprint('aluno', __name__)
+
+from . import views
