+++ main/aluno/views.py
@@ -1,40 +1,29 @@
-from flask import Flask, Response, request
-from flask_sqlalchemy import SQLAlchemy
-import json, os
-
-PASSWORD = os.environ.get('POSTGRES_PASSWORD')
-
-app = Flask(__name__)
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = True
-app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql://default:{}@ep-autumn-dream-20888543.us-east-1.postgres.vercel-storage.com:5432/verceldb'.format(PASSWORD)
-db = SQLAlchemy(app)
-
-class Usuario(db.Model):
-    id = db.Column(db.String(30), primary_key= True)
-    autorizado = db.Column(db.String(1))
-
-    def to_json(self):
-        return {"id": self.id, "autorizado": self.autorizado}
+import json
+from . import aluno
+from models import Usuario
+from response import gera_response
 
 
 # Selecionar Tudo
-@app.route("/usuarios", methods=["GET"])
+@aluno.route("/usuarios", methods=["GET"])
 def seleciona_usuarios():
     usuarios_objetos = Usuario.query.all()
     usuarios_json = [usuario.to_json() for usuario in usuarios_objetos]
 
     return gera_response(200, "usuarios", usuarios_json)
 
+
 # Selecionar Individual
-@app.route("/usuario/<id>", methods=["GET"])
+@aluno.route("/usuario/<id>", methods=["GET"])
 def seleciona_usuario(id):
     usuario_objeto = Usuario.query.filter_by(id=id).first()
     usuario_json = usuario_objeto.to_json()
 
     return gera_response(200, "usuario", usuario_json)
 
+
 # Atualizar
-@app.route("/change/<id>", methods=["PUT"])
+@aluno.route("/change/<id>", methods=["PUT"])
 def atualiza_usuario(id):
     usuario_objeto = Usuario.query.filter_by(id=id).first()
     body = request.get_json()
@@ -50,8 +39,9 @@ def atualiza_usuario(id):
         print('Erro', e)
         return gera_response(400, "usuario", {}, "Erro ao atualizar")
 
+
 # Deletar
-@app.route("/usuario/<id>", methods=["DELETE"])
+@aluno.route("/usuario/<id>", methods=["DELETE"])
 def deleta_usuario(id):
     usuario_objeto = Usuario.query.filter_by(id=id).first()
 
@@ -63,21 +53,12 @@ def deleta_usuario(id):
         print('Erro', e)
         return gera_response(400, "usuario", {}, "Erro ao deletar")
 
-@app.route("/")
+
+@aluno.route("/")
 def home():
     return "Hello World, from Flask!"
 
-@app.route("/login")
+
+@aluno.route("/login")
 def login():
     return "Login realizado"
-
-
-# JSON Return
-def gera_response(status, nome_do_conteudo, conteudo, mensagem=False):
-    body = {}
-    body[nome_do_conteudo] = conteudo
-
-    if(mensagem):
-        body["mensagem"] = mensagem
-
-    return Response(json.dumps(body), status=status, mimetype="application/json")
