+++ vercel.json
@@ -1,14 +1,14 @@
 {
   "builds": [
     {
-      "src": "api/main.py",
+      "src": "main/main.py",
       "use": "@vercel/python"
     }
   ],
   "routes": [
     {
       "src": "/(.*)",
-      "dest": "api/main.py"
+      "dest": "main/main.py"
     }
   ]
 }
