+++ wsgi.py
@@ -1,0 +1,4 @@
+from main import app
+
+if __name__ == "__main__":
+    app.run()
